% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncr.R
\name{ncr_nearest}
\alias{ncr_nearest}
\title{Fetches the number of charge points  within a search radius}
\usage{
ncr_nearest(lat, long)
}
\arguments{
\item{lat}{latitude in decimal degrees}

\item{long}{longitude in decimal degrees}
}
\description{
Fetches the number of charge points  within a search radius
}
\examples{
ncr_nearest(51.545581, -0.077301)
}
